0.0.2
The image `dharitri/sdk-rust-contract-builder:v0.0.1` is automatically created in the docker-hub if we make a release of it in github.

By using this image, two rust repos were built:
1. `drt-sc-sovereign`
2. `drt-rs-contracts`

Note: The previous commit's hash of these two rust repos have been included in this repo, maybe we can include the initial commits commit hash of these two rust repos in this drt-sdk-rust-contract-builder repo.